+++ bin/mutator-sensu-go-into-v1.rb
@@ -1,0 +1,13 @@
+#!/usr/bin/env ruby
+
+require 'sensu-mutator'
+
+class Helper < Sensu::Mutator
+
+  def mutate
+    new_event = self.map_go_event_into_v1
+    self.event= new_event
+  end
+
+end
+